README revamp
I haven't gone back and made it nice and pretty yet, but for now it has all the functional details to get an inexperienced user up and running. If I can think of a good picture/GIF to use for the feature I will add in everything else as well.